% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosPdf.R
\name{CosPdf}
\alias{CosPdf}
\title{CosPdf}
\usage{
CosPdf(y, a, b, N, Chf)
}
\arguments{
\item{y}{vector of observations}

\item{a}{the lower limit of the truncation interval}

\item{b}{the upper limit of the truncation interval}

\item{N}{the number of cos term for summation}

\item{Chf}{the characteristic function}
}
\value{
The approximated probability density of x
}
\description{
Using the cos method to restore the probability density of a function
}
\references{
Fang F. and Oosterlee C.W. 2008. "A Novel Pricing Method for European Options Based on Fourier-Cosine Series Expansions", Siam Journal on Scientific Computing. 31(2): 826-848. doi: 10.1137/080718061.
}
