% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosPlot.R
\name{CosPlot}
\alias{CosPlot}
\title{CosPlot}
\usage{
CosPlot(data, x, y)
}
\arguments{
\item{data}{a tibble， which contains x and y}

\item{x}{x}

\item{y}{y, which is the pdf returned from CosPdf}
}
\value{
A ggplot figure of the probability density function
}
\description{
Plot the p.d.f function for the univariate distribution with x and y
}
\examples{
N <- 64
y <- seq(-5, 5, by = 0.1)
a <- -6.0
b <- 6.0
NormChf <- function(u) {mu = 0;sigma = 1;return(exp(1i * mu * u - 0.5 * sigma ^ 2 * u ^ 2))}
cosnorm <- CosPdf(y, a, b, N, NormChf)
require("tibble")
cosnorm <- tibble(x=y,y=cosnorm)
nor <- CosPlot(data=cosnorm,x=x,y=y)
nor
}
