% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.r
\name{event_history_monthly}
\alias{event_history_monthly}
\title{Generate monthly event history dataset for a specified sample and date range}
\usage{
event_history_monthly(
  .start = "2023-01-01",
  .end = Sys.Date(),
  .sample = cosr::expr_valid_regs,
  .fn = cosr::resource_counts_at_date,
  .fileout
)
}
\arguments{
\item{.start}{The start date for the time series (default: "2023-01-01").}

\item{.end}{The end date for the time series (default: \code{Sys.Date()})}

\item{.sample}{The sample of registrations to analyze (e.g., \code{cosr::expr_valid_regs}).}

\item{.fn}{The function to apply for computing event counts at each cutoff date (default: \code{cosr::resource_counts_at_date}).}

\item{.fileout}{The file path for the output Parquet file (e.g., "data/event_history_monthly.parquet").}
}
\value{
A Parquet file containing the monthly event history dataset
}
\description{
This function generates a monthly time series dataset of event counts (e.g., OSR, LOS) for a specified sample and date range. It applies a user-defined function to compute event counts at each monthly cutoff date, combines the results into a single data frame, and saves it as a Parquet file.
}
