% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\docType{data}
\name{expr_valid_regs}
\alias{expr_valid_regs}
\title{Common filter expression for valid registrations}
\format{
An rlang expression list that can be spliced into dplyr::filter()
}
\usage{
expr_valid_regs
}
\description{
Defines "valid" Open Science Registrations (OSRs) as those that are:
\itemize{
\item Public (is_public == TRUE)
\item Not deprecated (is.na(deleted))
\item Have a registration date (!is.na(registered_date))
\item Accepted by moderation (moderation_state == "accepted")
\item Not spam (spam_status != 2 or is.na(spam_status))
}
}
\details{
This expression can be used with the \code{sample} parameter in most query
functions to filter the base set of registrations.
}
\examples{
\dontrun{
# Use in a query
valid_regs <- open_parquet(tbl = "osf_abstractnode") |>
  dplyr::filter(type == "osf.registration", !!!expr_valid_regs)

# Pass to a function
assign_attributes(sample = expr_valid_regs)
}

}
\keyword{datasets}
