% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.r
\name{get_funder_metadata}
\alias{get_funder_metadata}
\title{Get funder metadata for all OSF resources}
\usage{
get_funder_metadata(cache = TRUE)
}
\arguments{
\item{cache}{Logical. Should the results be cached? Defaults to \code{TRUE}.}
}
\value{
If \code{cache = TRUE}, returns the file path to the cached Parquet file.  If \code{cache = FALSE}, returns a tibble with funder metadata for all OSF resources.
}
\description{
Get funder metadata for all OSF resources
}
