% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.r
\name{get_registration_affiliated}
\alias{get_registration_affiliated}
\title{Get binary OSFI affiliation status for registrations}
\usage{
get_registration_affiliated(..., .simplify = TRUE)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=get_registration_institution]{get_registration_institution()}}}

\item{.simplify}{Logical. Defaults to TRUE}
}
\value{
A data frame with columns: node_id, affiliated
}
\description{
Convenience wrapper around \code{\link[=get_registration_institution]{get_registration_institution()}} that returns
simplified affiliated status.
}
