% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.r
\name{get_registration_funded}
\alias{get_registration_funded}
\title{Get binary funded status for registrations}
\usage{
get_registration_funded(..., .simplify = TRUE)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=get_registration_funder]{get_registration_funder()}}}

\item{.simplify}{Logical. Defaults to TRUE (always simplified)}
}
\value{
A data frame with columns: node_id, funded
}
\description{
Convenience wrapper around \code{\link[=get_registration_funder]{get_registration_funder()}} that returns
simplified "Funded"/"Unfunded" status.
}
