% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.r
\name{get_registration_provider}
\alias{get_registration_provider}
\title{Get provider information for registrations}
\usage{
get_registration_provider(
  sample = NULL,
  nodetype = "registration",
  lazy = TRUE
)
}
\arguments{
\item{sample}{Filter expression for sampling (default: NULL)}

\item{nodetype}{Type of node (default: "registration")}

\item{lazy}{Logical. Return lazy query? (default: TRUE)}
}
\value{
A data frame with columns: node_id, provider
}
\description{
Retrieves the OSF provider name for each registration.
}
