% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.r
\name{get_registration_subject}
\alias{get_registration_subject}
\title{Get subject classifications for registrations}
\usage{
get_registration_subject(sample = NULL, parents_only = FALSE, lazy = TRUE)
}
\arguments{
\item{sample}{Filter expression for sampling (default: NULL)}

\item{parents_only}{Logical. Return only parent subjects? (default: FALSE)}

\item{lazy}{Logical. Return lazy query? (default: TRUE)}
}
\value{
A data frame with columns: node_id, subject
}
\description{
Retrieves research subject classifications assigned to registrations.
}
