% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.r
\name{get_registration_subject_parent}
\alias{get_registration_subject_parent}
\title{Get parent subject classifications for registrations}
\usage{
get_registration_subject_parent(..., .parents_only = TRUE)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=get_registration_subject]{get_registration_subject()}}}

\item{.parents_only}{Logical. Defaults to TRUE}
}
\value{
A data frame with columns: node_id, subject_parent
}
\description{
Convenience wrapper around \code{\link[=get_registration_subject]{get_registration_subject()}} that returns
only parent (top-level) subjects.
}
