% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.r
\name{get_registration_template}
\alias{get_registration_template}
\title{Get registration template (schema) information}
\usage{
get_registration_template(sample = NULL, lazy = TRUE)
}
\arguments{
\item{sample}{Filter expression for sampling (default: NULL)}

\item{lazy}{Logical. Return lazy query? (default: TRUE)}
}
\value{
A data frame with columns: node_id, template
}
\description{
Retrieves the registration template/schema name used for each registration.
}
