% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.r
\name{los_summarizer}
\alias{los_summarizer}
\title{Generic summarizer for LOS time series data'}
\usage{
los_summarizer(tbl, lazy = TRUE, threshold = 0, ...)
}
\arguments{
\item{tbl}{The input time series dataset (e.g., \code{data/los_ts.parquet})}

\item{lazy}{Whether to return a lazy query (default: TRUE) or collect results into a data frame (default: FALSE)}

\item{threshold}{A numeric value between 0 and 1 specifying the minimum proportion of total OSRs required for a subgroup to be included in the summary (default: 0, meaning no threshold filter is applied)}

\item{...}{One or more unquoted expressions specifying the subgroup variables to summarize by (e.g., \code{institution}, \code{funded}, \code{provider}, \code{subject}, etc.)}
}
\value{
An (un)lazy data frame containing summary statistics for each subgroup defined by the specified variables, including counts and proportions of OSRs, resources, outputs, outcomes, and LOS.
}
\description{
This function takes a time series dataset (e.g., \code{data/los_ts.parquet}) and computes summary statistics (e.g., counts, proportions) for each subgroup defined by the specified variables.  It also applies an optional threshold filter to include only subgroups with a minimum share of the total available OSRs in the dataset.
}
