% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.r
\name{los_summary_by_attributes_all}
\alias{los_summary_by_attributes_all}
\title{Wrapper to compute time series summaries by multiple subgroup variables}
\usage{
los_summary_by_attributes_all(
  subgroup_vars = cosr::SUPPORTED_ATTRIBUTES,
  los_ts_path
)
}
\description{
Wrapper to compute time series summaries by multiple subgroup variables
}
