% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{open_parquet}
\alias{open_parquet}
\title{Open an OSF Parquet dataset}
\usage{
open_parquet(dir = Sys.getenv("PARQUET_DIR"), tbl, duck = TRUE)
}
\arguments{
\item{dir}{Path to directory containing Parquet files. Defaults to
\code{Sys.getenv("PARQUET_DIR")}, which can be set in your \code{.Renviron} file.}

\item{tbl}{Name of Parquet file (i.e., OSF database table name without
the .parquet extension)}

\item{duck}{Logical. Should the dataset be converted to a DuckDB dataset
using \code{arrow::to_duckdb()}? Defaults to \code{TRUE} for better query performance.}
}
\value{
An Arrow dataset or DuckDB connection
}
\description{
Opens a Parquet file from the OSF database directory and optionally
converts it to a DuckDB connection for efficient querying.
}
\details{
The Parquet directory path is read from the \code{PARQUET_DIR} environment
variable, which should be set in your \code{.Renviron} file.
}
\examples{
\dontrun{
# Open OSF abstractnode table
nodes <- open_parquet(tbl = "osf_abstractnode")

# Open as Arrow dataset (not DuckDB)
nodes_arrow <- open_parquet(tbl = "osf_abstractnode", duck = FALSE)
}

}
