% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.r
\docType{data}
\name{REPORTING_LABELS}
\alias{REPORTING_LABELS}
\title{Neat labels for reporting
A named vector of labels for reporting metrics. The names correspond to metric identifiers, and the values are human-readable labels that can be used in tables and visualizations.}
\format{
An object of class \code{character} of length 5.
}
\usage{
REPORTING_LABELS
}
\description{
Neat labels for reporting
A named vector of labels for reporting metrics. The names correspond to metric identifiers, and the values are human-readable labels that can be used in tables and visualizations.
}
\keyword{datasets}
