% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes.r
\name{resource_counts_at_date}
\alias{resource_counts_at_date}
\title{Count resources and LOS status at cutoff date}
\usage{
resource_counts_at_date(
  sample = cosr::expr_valid_regs,
  cutoff = Sys.Date(),
  lazy = TRUE
)
}
\arguments{
\item{sample}{The sample of registrations to analyze as an \code{rlang} expression to be defused and evaluated in the context of the \code{osf_abstractnode} table.  This defaults to \code{cosr::expr_valid_regs}.}

\item{cutoff}{The cutoff date for including resource history events (default: \code{Sys.Date()}).  Only resources linked or removed on or before this date will be included in the results.}

\item{lazy}{Whether to return a lazy query (default: TRUE) or collect results into a data frame (default: FALSE)}
}
\description{
This function computes summary statistics for registrations at a specified cutoff date, including counts of linked and removed resources (outputs and outcomes) and LOS status. It can be used to generate time series summaries of LOS by various attributes.
}
