% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{set_timespan}
\alias{set_timespan}
\title{Generate sequence of monthly cutoff dates for time series analysis}
\usage{
set_timespan(start = "2023-01-01", end = Sys.Date(), delta = "month")
}
\description{
Generate sequence of monthly cutoff dates for time series analysis
}
