% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.r
\docType{data}
\name{SUPPORTED_ATTRIBUTES}
\alias{SUPPORTED_ATTRIBUTES}
\title{Attributes currently supported for summarization and subgroup analyses}
\format{
An object of class \code{character} of length 8.
}
\usage{
SUPPORTED_ATTRIBUTES
}
\description{
Attributes currently supported for summarization and subgroup analyses
}
\keyword{datasets}
