% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.r
\name{assign_attributes}
\alias{assign_attributes}
\title{Assign current attributes (subgroup status for all OSRs}
\usage{
assign_attributes(sample = cosr::expr_valid_regs)
}
\arguments{
\item{sample}{A dplyr filter expression to apply to the base table of OSRs (e.g., \code{dplyr::filter(created >= "2024-01-01")}).  The default is \code{cosr::expr_valid_regs}, which captures all "valid" registrations.}
}
\description{
This function is designed to be run once per time period (e.g., monthly) to capture current attributes/ subgroup memberships for all OSRs.  The results are cached as Parquet files in the \verb{data/} directory for use in downstream analyses.
}
