% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{collector}
\alias{collector}
\title{Collect or return lazy dataset}
\usage{
collector(x, .lazy = TRUE)
}
\arguments{
\item{x}{A dataset (e.g., Arrow or DuckDB)}

\item{.lazy}{Logical. Should the dataset be returned in its lazy form (i.e., without collecting into memory)?  Defaults to \code{TRUE}.}
}
\description{
Collect or return lazy dataset
}
