% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.r
\docType{data}
\name{expr_valid_regs}
\alias{expr_valid_regs}
\title{The \code{samples} module contains predefined sample expressions and
utility functions for generating time spans.}
\format{
An object of class \code{list} of length 4.
}
\usage{
expr_valid_regs
}
\description{
The \code{samples} module contains predefined sample expressions and
utility functions for generating time spans.
}
\keyword{datasets}
