% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.r
\docType{data}
\name{expr_valid_regs}
\alias{expr_valid_regs}
\title{Common \code{rlang} expressions for filtering and summarizing datasets
Define valid (OSR) registrations}
\format{
An object of class \code{list} of length 5.
}
\usage{
expr_valid_regs
}
\description{
Common \code{rlang} expressions for filtering and summarizing datasets
Define valid (OSR) registrations
}
\keyword{datasets}
