% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes.r
\name{get_resource_history}
\alias{get_resource_history}
\title{Get resource linkage history}
\usage{
get_resource_history(
  sample = cosr::expr_valid_regs,
  cutoff = Sys.Date(),
  lazy = TRUE
)
}
\arguments{
\item{sample}{The sample of registrations to analyze as an \code{rlang} expression to be defused and evaluated in the context of the \code{osf_abstractnode} table.  This defaults to \code{cosr::expr_valid_regs}.}

\item{cutoff}{The cutoff date for including resource history events (default: \code{Sys.Date()}).  Only resources linked or removed on or before this date will be included in the results.}

\item{lazy}{Whether to return a lazy query (default: TRUE) or collect results into a data frame (default: FALSE)}
}
\description{
This function retrieves the history of resources (e.g., data, code, materials, papers, supplements) linked and removed from registrations up until a specified cutoff date. It combines data from multiple tables to create a comprehensive dataset of resource history for each registration.
}
