% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.r
\name{los_summary_by_attribute}
\alias{los_summary_by_attribute}
\title{Compute time series summaries by attribute}
\usage{
los_summary_by_attribute(subgroup_var, los_ts_path = "data/los_ts.parquet")
}
\arguments{
\item{subgroup_var}{The attribute to summarize by as a string (e.g., "institution", "funder", "funded", "provider", "subject", etc.)}

\item{los_ts_path}{The file path to the LOS time series dataset (e.g., "data/los_ts.parquet")}
}
\value{
An in-memory data frame containing summary statistics for each subgroup defined by the specified attribute variable
}
\description{
This function computes summary statistics for LOS time series data by specified attribute (or subgroup) variables. It loads the time series dataset, joins it with attribute data, and applies the \code{los_summarizer} function to compute summaries for each subgroup.
}
