% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.r
\name{open_parquet}
\alias{open_parquet}
\title{Open an OSF Parquet dataset}
\usage{
open_parquet(dir = Sys.getenv("PARQUET_DIR"), tbl, duck = TRUE)
}
\arguments{
\item{dir}{Path to directory containing Parquet files.  Defaults \code{Sys.getenv("PARQUET_DIR")}, which can be set in your \code{.Renviron} file.}

\item{tbl}{Name of Parquet file (i.e., OSF database table)}

\item{duck}{Logical. Should the dataset be converted to a DuckDB dataset using \code{duckdb::as_duckdb()}?  Defaults to \code{TRUE}.}
}
\description{
Open an OSF Parquet dataset
}
