% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.r
\docType{data}
\name{SUPPORTED_ATTRIBUTES}
\alias{SUPPORTED_ATTRIBUTES}
\title{Attributes currently supported for summarization and subgroup analyses}
\format{
Character vector of length 8
}
\usage{
SUPPORTED_ATTRIBUTES
}
\description{
This constant defines the 8 attribute types that can be used for
subgroup analysis in LOS workflows.
}
\keyword{datasets}
